# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Calder API",
    "version": "1.0.0",
    "description": "Email infrastructure: send, track, verify. Test keys simulate, never deliver.",
    "contact": { "name": "Calder", "url": "https://calder.click" }
  },
  "servers": [{ "url": "https://api.calder.click" }],
  "security": [{ "bearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A Calder secret key: calder_sk_test_<random> or calder_sk_live_<random>. The raw secret is shown only at creation or rotation."
      }
    },
    "schemas": {
...
      "CreateApiKey": {
        "type": "object",
        "required": ["name", "env"],
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "env": { "type": "string", "enum": ["test", "live"] },
          "scope": { "type": "string", "enum": ["full", "send", "read"], "default": "full" },
          "expires_at": {
            "type": ["string", "null"],
            "format": "date-time",
            "description": "Future expiration. Null or omission means no expiration."
          }
        }
      },
      "RotateApiKey": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "maxLength": 100 },
          "scope": { "type": "string", "enum": ["full", "send", "read"] },
          "expires_at": {
            "type": ["string", "null"],
            "format": "date-time",
            "description": "Future expiration; omission inherits the old key's expiry."
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/v1/keys": {
      "get": {
        "summary": "List API keys (no secrets)",
        "parameters": [
          { "name": "cursor", "in": "query", "schema": { "type": "string", "maxLength": 200 } },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
          }
        ],
        "responses": { "200": { "description": "Keys and cursor pagination" } }
      },
      "post": {
        "summary": "Create a key (secret shown once)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateApiKey" } }
          }
        },
        "responses": { "201": { "description": "Key + secret" } }
      }
    },
    "/v1/keys/{id}/rotate": {
      "post": {
        "summary": "Atomically rotate a key (old key is revoked)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RotateApiKey" } }
          }
        },
        "responses": { "201": { "description": "Replacement key + secret shown once" } }
      }
    },
    "/v1/keys/{id}/revoke": {
      "post": {
        "summary": "Revoke a key",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Key" } }
      }
    },
...
    "/v1/domains": {
      "get": {
        "summary": "List domains",
        "parameters": [
          { "name": "cursor", "in": "query", "schema": { "type": "string", "maxLength": 200 } },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
          }
        ],
        "responses": { "200": { "description": "Domains and cursor pagination" } }
      },
      "post": {
        "summary": "Add a domain (returns DNS records)",
        "responses": { "201": { "description": "Domain" } }
      }
    },
    "/v1/domains/{id}": {
      "delete": {
        "summary": "Remove a domain and its DNS challenge",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Deleted" } }
      }
    },
    "/v1/domains/{id}/token": {
      "post": {
        "summary": "Re-issue the DNS verification token",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Challenge (host + value)" },
          "409": { "description": "Already verified" }
        }
      }
    },
    "/v1/domains/{id}/verify": {
      "post": {
        "summary": "Check DNS and verify",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Domain" } }
      }
    },
    "/v1/domains/{id}/ses/link": {
      "post": {
        "summary": "Attach a verified SES identity (returns DKIM + SPF records)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Identity pending; add the returned records" },
          "409": { "description": "Verify DNS ownership first" },
          "429": { "description": "SES throttling, retry shortly" },
          "502": { "description": "SES identity creation failed" }
        }
      }
    },
    "/v1/domains/{id}/ses/refresh": {
      "post": {
        "summary": "Refresh SES identity and DKIM status",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Status snapshot" } }
      }
    },
...
  }
}